Fix TransactionResponse, move duplicated code to Response class.
