+++ lib/Core/Response/CaseResponse.php
@@ -274,20 +274,6 @@ class CaseResponse extends Response
     }
 
     /**
-     * Set the error
-     *
-     * @param int    $httpCode The response code
-     * @param string $error    The response
-     *
-     * @return void
-     */
-    public function setError($httpCode, $error)
-    {
-        $this->setIsError(true);
-        $this->setErrorMessage($error);
-    }
-
-    /**
      * Get the eligible for guarantee
      *
      * @return bool
@@ -618,51 +604,6 @@ class CaseResponse extends Response
     {
         $this->updatedAt = $updatedAt;
     }
-
-    /**
-     * Is the response in error
-     *
-     * @return bool
-     */
-    public function isError()
-    {
-        return $this->isError;
-    }
-
-    /**
-     * Set the error for the response object
-     *
-     * @param bool $isError The error state
-     *
-     * @return void
-     */
-    public function setIsError($isError)
-    {
-        $this->isError = $isError;
-    }
-
-    /**
-     * Get the error message
-     *
-     * @return string
-     */
-    public function getErrorMessage()
-    {
-        return $this->errorMessage;
-    }
-
-    /**
-     * Set the error message
-     *
-     * @param string $errorMessage The error message
-     *
-     * @return void
-     */
-    public function setErrorMessage($errorMessage)
-    {
-        $this->errorMessage = $errorMessage;
-    }
-
 
     /**
      * Set the investigation id to the case id
