+++ lib/Core/Response/FulfillmentBulkResponse.php
@@ -73,51 +73,6 @@ class FulfillmentBulkResponse extends Response
     }
 
     /**
-     * Is the response in error
-     *
-     * @return bool
-     */
-    public function isError()
-    {
-        return $this->isError;
-    }
-
-    /**
-     * Set the response error
-     *
-     * @param bool $isError If response is in error
-     *
-     * @return void
-     */
-    public function setIsError($isError)
-    {
-        $this->isError = $isError;
-    }
-
-    /**
-     * Get the error message
-     *
-     * @return string
-     */
-    public function getErrorMessage()
-    {
-        return $this->errorMessage;
-    }
-
-    /**
-     * Set the error message
-     *
-     * @param string $errorMessage The error message
-     *
-     * @return void
-     */
-    public function setErrorMessage($errorMessage)
-    {
-        $this->errorMessage = $errorMessage;
-    }
-
-
-    /**
      * Get the response objects
      *
      * @return array
@@ -143,17 +98,4 @@ class FulfillmentBulkResponse extends Response
         }
     }
 
-    /**
-     * Set the error
-     *
-     * @param int    $httpCode The response code
-     * @param string $error    The response
-     *
-     * @return void
-     */
-    public function setError($httpCode, $error)
-    {
-        $this->setIsError(true);
-        $this->setErrorMessage($error);
-    }
 }
